feat/#61 - pk Long형으로 바꿈

+++ neulbomgil-backend/src/main/java/com/gooroomees/neulbomgil_backend/domain/chat/controller/ChatWebSocketController.java
@@ -19,7 +19,7 @@ public class ChatWebSocketController {
     //pub/chatrooms/{roomId}/messages
     @MessageMapping("/chatrooms/{roomId}/messages")
     public void sendMessage(
-            @DestinationVariable Integer roomId,
+            @DestinationVariable Long roomId,
             ChatRequestDto requestDto
     ){
         ChatResponseDto responseDto = chatService.saveMessage(roomId,requestDto);
